rulesData: Break tableView:acceptDrop:... into type-specific methods.
Sort of necessary to help thinking about the drop types and the
operations needed for each.

Selector pastes can only be dropped on rules, rule pastes can only be
dropped above rules (i.e., you can't merge rules).

Mixed-type pastes are supposed to be impossible, but there's a check
for that anyway.