RevisionTracker: remove extension; just retain an NSUndoManager.
Not yet sure why, but allocating the QRevisionTracker instance from the
extension seems to be causing an early release, so that needs working
out at some point.

In the meantime, this appears to avoid the issue: allocating it
directly.

No point retaining the document itself either, so this just takes an
undo manager and keeps a weak reference to it.

(Also fixed the file header.)